requirements.txt
numpy
scipy
matplotlib
seaborn
torch
omegaconf
hydra-core